Added AST Batching support
