+++ rules-java/api/src/main/java/org/jboss/windup/rules/apps/java/scan/provider/AnalyzeJavaFilesRuleProvider.java
@@ -1,18 +1,24 @@
 package org.jboss.windup.rules.apps.java.scan.provider;
 
-import java.io.File;
+import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.util.HashMap;
 import java.util.HashSet;
+import java.util.List;
 import java.util.Map;
 import java.util.Set;
+import java.util.TreeMap;
+import java.util.TreeSet;
+import java.util.concurrent.atomic.AtomicInteger;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
 import javax.inject.Inject;
 
 import org.jboss.windup.ast.java.ASTProcessor;
+import org.jboss.windup.ast.java.BatchASTListener;
+import org.jboss.windup.ast.java.BatchASTProcessor;
 import org.jboss.windup.ast.java.data.ClassReference;
-import org.jboss.windup.ast.java.data.ClassReferences;
 import org.jboss.windup.ast.java.data.TypeReferenceLocation;
 import org.jboss.windup.ast.java.data.annotations.AnnotationArrayValue;
 import org.jboss.windup.ast.java.data.annotations.AnnotationClassReference;
@@ -21,13 +27,8 @@ import org.jboss.windup.ast.java.data.annotations.AnnotationValue;
 import org.jboss.windup.config.AbstractRuleProvider;
 import org.jboss.windup.config.GraphRewrite;
 import org.jboss.windup.config.metadata.MetadataBuilder;
-import org.jboss.windup.config.operation.Commit;
 import org.jboss.windup.config.operation.GraphOperation;
-import org.jboss.windup.config.operation.Iteration;
-import org.jboss.windup.config.operation.IterationProgress;
-import org.jboss.windup.config.operation.iteration.AbstractIterationOperation;
 import org.jboss.windup.config.phase.InitialAnalysisPhase;
-import org.jboss.windup.config.query.Query;
 import org.jboss.windup.graph.GraphContext;
 import org.jboss.windup.graph.model.resource.FileModel;
 import org.jboss.windup.graph.service.GraphService;
@@ -56,9 +57,10 @@ import org.ocpsoft.rewrite.context.EvaluationContext;
  */
 public class AnalyzeJavaFilesRuleProvider extends AbstractRuleProvider
 {
+    public static final int COMMIT_INTERVAL = 750;
+    public static final int LOG_INTERVAL = 250;
     private static Logger LOG = Logging.get(AnalyzeJavaFilesRuleProvider.class);
 
-    private ASTProcessor processor;
     @Inject
     private WindupWildcardImportResolver importResolver;
 
@@ -75,115 +77,184 @@ public class AnalyzeJavaFilesRuleProvider extends AbstractRuleProvider
     {
         return ConfigurationBuilder.begin()
             .addRule()
-            .when(Query.fromType(JavaSourceFileModel.class))
-            .perform(
-                    Iteration.over().perform(
-                        new ParseSourceOperation()
-                            .and(IterationProgress.monitoring("Analyzed Java File: ", 250))
-                            .and(Commit.every(10))
-                    ).endIteration()
-                    .and(new ClearClasspathCache())
-            );
+            .perform(new ParseSourceOperation());
+
+
+                                    //.and(IterationProgress.monitoring("Analyzed Java File: ", 250))
+                                    //.and(Commit.every(10))
     }
     // @formatter:on
 
-    private final class ClearClasspathCache extends GraphOperation
+    private final class ParseSourceOperation extends GraphOperation
     {
-        @Override
-        public void perform(GraphRewrite event, EvaluationContext context)
-        {
-            processor = null;
-        }
-    }
-
-    private final class ParseSourceOperation extends AbstractIterationOperation<JavaSourceFileModel>
-    {
-        public void perform(GraphRewrite event, EvaluationContext context, JavaSourceFileModel payload)
+        public void perform(final GraphRewrite event, EvaluationContext context)
         {
             ExecutionStatistics.get().begin("AnalyzeJavaFilesRuleProvider.analyzeFile");
             try
             {
                 WindupJavaConfigurationService windupJavaConfigurationService = new WindupJavaConfigurationService(
                             event.getGraphContext());
-                if (!windupJavaConfigurationService.shouldScanPackage(payload.getPackageName()))
-                {
-                    // should not analyze this one, skip it
-                    return;
-                }
 
-                GraphService<JavaSourceFileModel> service = new GraphService<JavaSourceFileModel>(event.getGraphContext(), JavaSourceFileModel.class);
-                if (processor == null)
+                GraphService<JavaSourceFileModel> service = new GraphService<>(event.getGraphContext(), JavaSourceFileModel.class);
+                Iterable<JavaSourceFileModel> allJavaSourceModels = service.findAll();
+
+                final Set<Path> allSourceFiles = new TreeSet<>();
+                final Map<Path, JavaSourceFileModel> sourcePathToFileModel = new TreeMap<>();
+
+                Set<String> sourcePaths = new HashSet<>();
+                for (JavaSourceFileModel javaFile : allJavaSourceModels)
                 {
-                    Iterable<JavaSourceFileModel> allJavaSourceModels = service.findAll();
-                    Set<String> sourcePaths = new HashSet<>();
-                    for (JavaSourceFileModel javaFile : allJavaSourceModels)
+                    FileModel rootSourceFolder = javaFile.getRootSourceFolder();
+                    if (rootSourceFolder != null)
                     {
-                        FileModel rootSourceFolder = javaFile.getRootSourceFolder();
-                        if (rootSourceFolder != null)
-                        {
-                            sourcePaths.add(rootSourceFolder.getFilePath());
-                        }
+                        sourcePaths.add(rootSourceFolder.getFilePath());
                     }
 
-                    GraphService<JarArchiveModel> libraryService = new GraphService<JarArchiveModel>(event.getGraphContext(), JarArchiveModel.class);
-
-                    Iterable<JarArchiveModel> libraries = libraryService.findAll();
-                    Set<String> libraryPaths = new HashSet<>();
-                    for (JarArchiveModel library : libraries)
+                    if (windupJavaConfigurationService.shouldScanPackage(javaFile.getPackageName()))
                     {
-                        if (library.getUnzippedDirectory() != null)
-                        {
-                            libraryPaths.add(library.getUnzippedDirectory().getFilePath());
-                        }
-                        else
-                        {
-                            libraryPaths.add(library.getFilePath());
-                        }
+                        Path path = Paths.get(javaFile.getFilePath());
+                        allSourceFiles.add(path);
+                        sourcePathToFileModel.put(path, javaFile);
                     }
-                    processor = new ASTProcessor(importResolver, libraryPaths, sourcePaths);
                 }
-                File sourceFile = payload.asFile();
 
-                ExecutionStatistics.get().begin("AnalyzeJavaFilesRuleProvider.parseFile");
+                GraphService<JarArchiveModel> libraryService = new GraphService<JarArchiveModel>(event.getGraphContext(), JarArchiveModel.class);
+
+                Iterable<JarArchiveModel> libraries = libraryService.findAll();
+                Set<String> libraryPaths = new HashSet<>();
+                for (JarArchiveModel library : libraries)
+                {
+                    if (library.getUnzippedDirectory() != null)
+                    {
+                        libraryPaths.add(library.getUnzippedDirectory().getFilePath());
+                    }
+                    else
+                    {
+                        libraryPaths.add(library.getFilePath());
+                    }
+                }
+                ExecutionStatistics.get().begin("AnalyzeJavaFilesRuleProvider.parseFiles");
                 try
                 {
                     WindupWildcardImportResolver.setGraphContext(event.getGraphContext());
-                    ClassReferences references = processor.analyzeFile(sourceFile.toPath());
-                    TypeReferenceService typeReferenceService = new TypeReferenceService(event.getGraphContext());
-                    for (ClassReference reference : references.getReferences())
+                    final int totalToProcess = allSourceFiles.size();
+                    final AtomicInteger numberProcessed = new AtomicInteger(0);
+
+                    final Set<Path> processedPaths = new HashSet<>(allSourceFiles.size());
+                    BatchASTListener listener = new BatchASTListener()
                     {
-                        // we are always interested in types + anything that the TypeInterestFactory has registered
-                        if (reference.getLocation() == TypeReferenceLocation.TYPE
-                                    || TypeInterestFactory.matchesAny(reference.getQualifiedName(), reference.getLocation()))
+                        @Override
+                        public void processed(Path filePath, List<ClassReference> references)
                         {
-                            JavaTypeReferenceModel typeReference = typeReferenceService.createTypeReference(payload, reference.getLocation(),
-                                        reference.getLineNumber(), reference.getColumn(), reference.getLength(), reference.getQualifiedName(),
-                                        reference.getLine());
-                            if (reference instanceof AnnotationClassReference)
+                            processedPaths.add(filePath);
+                            processReferences(event.getGraphContext(), sourcePathToFileModel, filePath, references);
+
+                            numberProcessed.incrementAndGet();
+                            if (numberProcessed.get() % LOG_INTERVAL == 0)
                             {
-                                Map<String, AnnotationValue> annotationValues = ((AnnotationClassReference) reference).getAnnotationValues();
-                                addAnnotationValues(event.getGraphContext(), typeReference, annotationValues);
+                                LOG.info("Analyzed Java File: " + numberProcessed.get() + " / " + totalToProcess);
+                            }
+
+                            if (numberProcessed.get() % COMMIT_INTERVAL == 0)
+                            {
+                                event.getGraphContext().getGraph().getBaseGraph().commit();
+                            }
+                        }
+
+                        @Override
+                        public void failed(Path filePath, Throwable cause)
+                        {
+                            LOG.log(Level.WARNING, "Failed to process: " + filePath + " due to: " + cause.getMessage(), cause);
+                            ClassificationService classificationService = new ClassificationService(event.getGraphContext());
+                            JavaSourceFileModel sourceFileModel = sourcePathToFileModel.get(filePath);
+                            classificationService.attachClassification(sourceFileModel, JavaSourceFileModel.UNPARSEABLE_JAVA_CLASSIFICATION,
+                                        JavaSourceFileModel.UNPARSEABLE_JAVA_DESCRIPTION);
+                        }
+                    };
+
+                    Set<Path> filesToProcess = new TreeSet<>(allSourceFiles);
+                    BatchASTProcessor.analyze(listener, importResolver, libraryPaths, sourcePaths, filesToProcess);
+                    filesToProcess.removeAll(processedPaths);
+                    processedPaths.clear();
+
+                    if (!filesToProcess.isEmpty())
+                    {
+                        // try these one file at a time
+                        for (Path unprocessed : filesToProcess)
+                        {
+                            try
+                            {
+                                List<ClassReference> references = ASTProcessor.analyze(importResolver, libraryPaths, sourcePaths, unprocessed);
+                                processReferences(event.getGraphContext(), sourcePathToFileModel, unprocessed, references);
+                                processedPaths.add(unprocessed);
+                            }
+                            catch (Exception e)
+                            {
+                                LOG.log(Level.WARNING, "Failed to process: " + unprocessed + " due to: " + e.getMessage(), e);
+                                ClassificationService classificationService = new ClassificationService(event.getGraphContext());
+                                JavaSourceFileModel sourceFileModel = sourcePathToFileModel.get(unprocessed);
+                                classificationService.attachClassification(sourceFileModel, JavaSourceFileModel.UNPARSEABLE_JAVA_CLASSIFICATION,
+                                            JavaSourceFileModel.UNPARSEABLE_JAVA_DESCRIPTION);
                             }
                         }
                     }
-                    ExecutionStatistics.get().end("AnalyzeJavaFilesRuleProvider.parseFile");
+                    filesToProcess.removeAll(processedPaths);
+
+                    if (!filesToProcess.isEmpty())
+                    {
+                        ClassificationService classificationService = new ClassificationService(event.getGraphContext());
+
+                        StringBuilder message = new StringBuilder();
+                        message.append("Failed to process " + filesToProcess.size() + " files:\n");
+                        for (Path unprocessed : filesToProcess)
+                        {
+                            JavaSourceFileModel sourceFileModel = sourcePathToFileModel.get(unprocessed);
+                            message.append("\tFailed to process: " + unprocessed + "\n");
+                            classificationService.attachClassification(sourceFileModel, JavaSourceFileModel.UNPARSEABLE_JAVA_CLASSIFICATION,
+                                        JavaSourceFileModel.UNPARSEABLE_JAVA_DESCRIPTION);
+                        }
+                        LOG.warning(message.toString());
+                    }
+
+                    ExecutionStatistics.get().end("AnalyzeJavaFilesRuleProvider.parseFiles");
                 }
                 catch (Exception e)
                 {
-                    LOG.log(Level.WARNING, "Could not analyze java file: " + payload.getFilePath() + " due to: " + e.getMessage(), e);
-                    ClassificationService classificationService = new ClassificationService(event.getGraphContext());
-                    classificationService.attachClassification(payload, JavaSourceFileModel.UNPARSEABLE_JAVA_CLASSIFICATION,
-                                JavaSourceFileModel.UNPARSEABLE_JAVA_DESCRIPTION);
+                    LOG.log(Level.SEVERE, "Could not analyze java files: " + e.getMessage(), e);
                 }
                 finally
                 {
                     WindupWildcardImportResolver.setGraphContext(null);
                 }
-
             }
             finally
             {
                 ExecutionStatistics.get().end("AnalyzeJavaFilesRuleProvider.analyzeFile");
+            }
+        }
+
+        private void processReferences(GraphContext context, Map<Path, JavaSourceFileModel> sourcePathToFileModel, Path filePath,
+                    List<ClassReference> references)
+        {
+            TypeReferenceService typeReferenceService = new TypeReferenceService(context);
+            for (ClassReference reference : references)
+            {
+                // we are always interested in types + anything that the TypeInterestFactory has registered
+                if (reference.getLocation() == TypeReferenceLocation.TYPE
+                            || TypeInterestFactory.matchesAny(reference.getQualifiedName(), reference.getLocation()))
+                {
+                    JavaSourceFileModel javaSourceModel = sourcePathToFileModel.get(filePath);
+                    JavaTypeReferenceModel typeReference = typeReferenceService.createTypeReference(javaSourceModel,
+                                reference.getLocation(),
+                                reference.getLineNumber(), reference.getColumn(), reference.getLength(),
+                                reference.getQualifiedName(),
+                                reference.getLine());
+                    if (reference instanceof AnnotationClassReference)
+                    {
+                        Map<String, AnnotationValue> annotationValues = ((AnnotationClassReference) reference).getAnnotationValues();
+                        addAnnotationValues(context, typeReference, annotationValues);
+                    }
+                }
             }
         }
